Add regex and test case for Invoking with exception
