+++ FluentAssertionsToShouldlyConverter/FluentAssertionsToShouldlyConverter.csproj
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <RootNamespace>FluentAssertionsToShouldly</RootNamespace>
+  </PropertyGroup>
+
+</Project>
